+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v0.3beta6 2026-01-13
+- Added option to set charge limit defaults in settings.
 - Changed iot_class back to local_polling as it best describes SolarCharger communication with devices.
 
 ## v0.3beta5 2026-01-12